Font settings part 2.
After more and more tinkering rendering in terminal is more or less
good.  But I had to revert back to old (and pixel-perfect) font in i3;
especially window titles, where contrast can be low, titles wore
unreadable.  Still a lot to learn.

And this is one example of linux fallacy.  Many small tools, even
more settings, and sometimes it is impossible to make it all work together.